Move to org.apache.spark and make it follow the way of InterruptibleIterator.
